requirements.txt
# STORM Protocol Requirements

# Core async networking
dnspython>=2.3.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
pylint>=2.16.0
mypy>=0.991

# Optional: Performance monitoring
psutil>=5.9.0

# Security/encryption
cryptography>=42.0.0
